### 2025-08-04 #8705 @arfon
@aglucaci - thanks for your submission to JOSS. Could you please highlight the software contributions that make up the bulk of this project? We don't get a ton of Snakemake submissions so your help is appreciated here.

```
Language                     files          blank        comment           code
JSON                           173              0              0         250635
CSV                             49              0              0           9505
Brainfuck                        4             83             85            805
Python                           4            115            255            405
Snakemake                        1             76            129            289
Bourne Shell                     5             93             96            279
TeX                              1             29              0            275
Markdown                         2             94              0            231
YAML                             6             12             25             92
SVG                             34              0              0             34
SUM:                           279            502            590         262550
```
