### 2025-09-15 #8705 @aglucaci
Hello @Kevin-Mattheus-Moerman,

Yes, you are correct that Python, and more importantly the Snakemake file carries the main functionality of this submission.

I believe the .bf is misattributed. It actually stands for HyPhy Batch Language (HBL), which is a custom scripting language. More information about HBL can be found in the publication: https://academic.oup.com/mbe/article/37/1/295/5555420, along with a language reference here https://www.hyphy.org/hbl-reference/libv3/#table-of-contents.

Within the context of my submission - it is used for utility scripts that perform various functions.
