### 2025-12-23 #8705 @skanwal
Hello @hseabolt - are you available to review this submission for JOSS?
