### 2025-12-23 #8705 @hseabolt
> Hello [@hseabolt](https://github.com/hseabolt) - are you available to review this submission for JOSS?

Unfortunately I am not available to review at this time, sorry!
