### 2026-01-15 #8705 @stulacy
> 👋 [@stulacy](https://github.com/stulacy) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sorry for the delay, I was hoping I'd be able to say yes but I've got several big commitments until March. If you still need someone by then I'll be happy to help
