### 2026-01-21 #9872 @editorialbot
[REVIEW]: AOC: A Snakemake workflow for the characterization of natural selection in protein-coding genes
**Submitting author:** @aglucaci (<a href="http://orcid.org/0000-0002-4896-6088">Alexander Lucaci</a>)
**Repository:** https://github.com/aglucaci/Analysis-of-Orthologous-Collections
**Branch with paper.md** (empty if default branch):
**Version:** 1.12
