### 2026-01-30 #9872 @gavinmdouglas
Hi @skanwal, could you clarify whether all the required sections indicated in the software paper are indeed necessary? Most are missing from the manuscript, but I see these aren't listed in the other reviewer's checklist. Also, they don't seem to be present in these most recent JOSS publications: [https://doi.org/10.21105/joss.08389](https://doi.org/10.21105/joss.08389) and [https://doi.org/10.21105/joss.09343](https://doi.org/10.21105/joss.09343)
