### 2026-01-30 #9872 @gavinmdouglas
Also, @skanwal, your feedback on to what degree this project needs to be a collaboration between multiple developers would be appreciated. Clearly this is the project of just these two researchers (and one developer), as seen here: https://github.com/aglucaci/Analysis-of-Orthologous-Collections/graphs/contributors

I think that's totally fine, and personally have no issue with most of the development having been before it was pushed to be open on GitHub. However, I see that this means it doesn't fit the criteria in **Development history and open-source practice**. Does this mean that this tool cannot be published in JOSS? Again, I think that would be overly strict, but I want to make sure I'm following the required reviewer criteria.

Thanks!
