### 2026-02-02 #9872 @skanwal
Hi @arfon - any ideas why there different reviewer checklists being generated for sotware paper section?

Thanks @gavinmdouglas for raising these points. I agree that it's fine to have development happen outside of GitHub before pushing the work to the repository.

@arfon, are you also in agreement with these points?
