### 2026-02-06 #9872 @gavinmdouglas
Thanks, @skanwal, I hope @arfon will have a chance to speak on these points.

Just a quick note that I was able to install the software with conda, and I posted my initial review here: https://github.com/aglucaci/Analysis-of-Orthologous-Collections/issues/9
