### 2026-02-25 #9872 @gavinmdouglas
Hi @arfon, just pinging you again in the hopes you will be able to address my questions above. I wouldn't want the authors to spend a lot of time revising their manuscript and tool if they aren't allowed to publish in JOSS for not fulfilling the "Development history and open-source practice" criteria.

Also, are all specified sections in the review criteria actually necessary (see above)? It looks like these sections are missing from the most recent JOSS publications.
