### 2026-03-01 #9872 @aglucaci
Hello @gavinmdouglas @arfon and others from the JOSS team - I have shared my response to reviewer comments here: https://github.com/aglucaci/AOC/issues/9#issuecomment-3980309772. Please let me know if you have any questions, or if I need to make new modifications. A new PDF version has been generated: https://github.com/aglucaci/AOC/blob/main/JOSS/AOC.pdf.
