### 2026-03-04 #9872 @marcrr
Hello @skanwal , I'm a bit confused, how is this different from https://github.com/openjournals/joss-reviews/issues/8705#issuecomment-3685057517 ?
