### 2026-03-22 #9872 @skanwal
Hello @juanvillada - you can use the following command to generate your checklist:
