### 2026-03-23 #9872 @gavinmdouglas
Hello,

I finished reviewing the manuscript and discussing with the author (https://github.com/aglucaci/AOC/issues/9). I confirmed most of the review criteria, except for the below points.

On Development history and open-source practice (**but I was satisfied with the author's explanation regarding these points**):
- Development timeline
- Open development
- Collaborative effort

And these sections of the software paper are missing: Software Design, Research Impact Statement, and AI usage disclosure. I recommend the author (@aglucaci) add these sections, but since these are missing from the recent JOSS publications I saw, I wasn't sure how strict this format is. I will leave it to the editor @skanwal to decide on whether they will require those sections prior to publication!

Cheers,

Gavin
