### 2026-04-15 #9872 @skanwal
Thanks @gavinmdouglas. Can you please check off the following in the checklist if you're satisfied with authors response.

-  Development timeline
- Open development
- Collaborative effort

I agree and recommend adding the sections for Software Design, Research Impact Statement, and AI usage disclosure.
