### 2026-04-16 #9872 @gavinmdouglas
Sure, @skanwal - just done. And ok regarding those sections. @aglucaci it sounds like the sections Software Design, Research Impact Statement, and AI usage disclosure (even if no AI tools were used) are all required for your manuscript. I think these could be quite succinct and hit the specific points under each item described here: https://joss.readthedocs.io/en/latest/review_criteria.html
