### 2026-04-20 #9872 @juanvillada
Hi @aglucaci, I completed my review. Please see the comments here: https://github.com/aglucaci/AOC/issues/22. Thanks!
