### 2026-04-22 #9872 @juanvillada
Hi @skanwal, I have completed my review and my comments were satisfactorily answered by the author (See comments here: https://github.com/aglucaci/AOC/issues/22). Thanks!
