### 2026-05-01 #9872 @aglucaci
Hello @skanwal,

I have gone through each of the items that are reported as missing DOI's and they all appear to be good to me. I looked each one up on doi.org and here is my report for each one below.

❌ MISSING DOIs
- 10.1093/molbev/msi105 may be a valid DOI for title: Not So Different After All: A Comparison of Method...
Good!
- 10.1093/molbev/msl051 may be a valid DOI for title: Automated Phylogenetic Detection of Recombination ...
Good!
- 10.1093/molbev/msaa263 may be a valid DOI for title: Contrast-FEL-A Test for Differences in Selective P...
Good!
- 10.1093/molbev/msad150 may be a valid DOI for title: Evolutionary shortcuts via multi-nucleotide substi...
Good!
- 10.1038/s41398-022-02021-w may be a valid DOI for title: The evolution of BDNF is defined by strict purifyi...
Good!
- 10.1371/journal.pone.0248337 may be a valid DOI for title: Extra base hits: Widespread empirical support for ...
Good!
- 10.1093/molbev/msac061 may be a valid DOI for title: Selection analysis identifies clusters of unusual ...
Good!
- 10.1016/j.cell.2021.09.003 may be a valid DOI for title: The emergence and ongoing convergent evolution of ...
Good!
- 10.1093/molbev/msaa015 may be a valid DOI for title: IQ-TREE 2: New Models and Efficient Methods for Ph...
Good!
- 10.1093/molbev/mst030 may be a valid DOI for title: FUBAR: A Fast, Unconstrained Bayesian AppRoximatio...
Good!
- 10.1371/journal.pgen.1002764 may be a valid DOI for title: Detecting individual sites subject to episodic div...
Good!
- 10.1038/d41586-021-01174-w may be a valid DOI for title: Reactive, reproducible, collaborative: computation...
Good!
- 10.1093/nar/gkaa680 may be a valid DOI for title: DGINN, an automated and highly-flexible pipeline f...
Good!
- 10.1093/bioinformatics/btn313 may be a valid DOI for title: Spidermonkey: rapid detection of co-evolving sites...
Good!
- 10.1093/molbev/msv022 may be a valid DOI for title: Less is more: an adaptive branch-site random effec...
Good!
- 10.1093/molbev/msu400 may be a valid DOI for title: RELAX: detecting relaxed selection in a phylogenet...
Good!
- 10.1093/molbev/msaa037 may be a valid DOI for title: Synonymous Site-to-Site Substitution Rate Variatio...
Good!
- 10.1093/ve/vead019 may be a valid DOI for title: Natural selection differences detected in key prot...
Good!

All the best,
Alexander
