### 2026-05-08 #9872 @aglucaci
Hello @skanwal,

Here is the update.

Additional Author Tasks After Review is Complete
•	Double check authors and affiliations (including ORCIDs)
Done!

•	Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
Release Version is 1.6, link to it is here: https://github.com/aglucaci/AOC/releases/tag/1.6

•	Archive the release on Zenodo/figshare/etc and post the DOI here.
https://doi.org/10.48550/arXiv.2406.09522

•	Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
Done

•	Make sure that the license listed for the archive is the same as the software license.
Done

All the best,
Alexander
