### 2026-05-11 #9872 @aglucaci
Hello @skanwal @danielskatz,

Here is the updated version that I believe is appropriate for this purpose: https://doi.org/10.5281/zenodo.20115222.

I have uploaded the archived version 1.6 of AOC there.  Here is also a link to Zenodo: https://zenodo.org/records/20115222

All the best,
Alexander
